Trait pallet_statement::pallet::Config

pub trait Config: Config
where
    <Self as Config>::AccountId: From<AccountId>,
{
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: Inspect<Self::AccountId>;
    type StatementCost: Get<BalanceOf<Self>>;
    type ByteCost: Get<BalanceOf<Self>>;
    type MinAllowedStatements: Get<u32>;
    type MaxAllowedStatements: Get<u32>;
    type MinAllowedBytes: Get<u32>;
    type MaxAllowedBytes: Get<u32>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type Currency: Inspect<Self::AccountId>

The currency which is used to calculate account limits.

type StatementCost: Get<BalanceOf<Self>>

Min balance for priority statements.

type ByteCost: Get<BalanceOf<Self>>

Cost of data byte used for priority calculation.

type MinAllowedStatements: Get<u32>

Minimum number of statements allowed per account.

type MaxAllowedStatements: Get<u32>

Maximum number of statements allowed per account.

type MinAllowedBytes: Get<u32>

Minimum data bytes allowed per account.

type MaxAllowedBytes: Get<u32>

Maximum data bytes allowed per account.

Object Safety

This trait is not object safe.